examples/python/pyproject.toml
[project]
name = "wsclient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff>=0.12.4",
    "websocket-client>=1.8.0",
]
